Move partition synopsis into separate header
Move the partition synopsis into its own header, and remove
obsolete functions from the meta index interface.
