+++ libvast/src/detail/add_message_types.cpp
@@ -18,7 +18,6 @@
 #include "vast/command.hpp"
 #include "vast/config.hpp"
 #include "vast/expression.hpp"
-#include "vast/meta_index.hpp"
 #include "vast/operator.hpp"
 #include "vast/query_options.hpp"
 #include "vast/schema.hpp"
