+++ libvast/vast/meta_index.hpp
@@ -13,10 +13,12 @@
 
 #pragma once
 
+#include "vast/fwd.hpp"
+
 #include "vast/fbs/index.hpp"
 #include "vast/fbs/partition.hpp"
-#include "vast/fwd.hpp"
 #include "vast/ids.hpp"
+#include "vast/partition_synopsis.hpp"
 #include "vast/qualified_record_field.hpp"
 #include "vast/synopsis.hpp"
 #include "vast/time_synopsis.hpp"
@@ -45,43 +47,14 @@ pack(flatbuffers::FlatBufferBuilder& builder,
 
 } // namespace system
 
-/// Contains one synopsis per partition column.
-//  TODO: Move this into a separate `partition_synopsis.hpp` header.
-struct partition_synopsis {
-  /// Optimizes the partition synopsis contents for size.
-  void shrink();
-
-  void add(const table_slice& slice, const caf::settings& synopsis_options);
-
-  /// @returns A best-effort estimate of the amount of memory used by this
-  ///          synopsis.
-  size_t size_bytes() const;
-
-  /// Synopsis data structures for types.
-  std::unordered_map<type, synopsis_ptr> type_synopses_;
-
-  /// Synopsis data structures for individual columns.
-  std::unordered_map<qualified_record_field, synopsis_ptr> field_synopses_;
-};
-
 /// The meta index is the first data structure that queries hit. The result
 /// represents a list of candidate partition IDs that may contain the desired
 /// data. The meta index may return false positives but never false negatives.
 class meta_index {
 public:
-  /// Adds all data from a table slice belonging to a given partition to the
-  /// index.
-  /// @param slice The table slice to extract data from.
-  /// @param partition The partition ID that *slice* belongs to.
-  void add(const uuid& partition, const table_slice& slice);
-
   /// Adds new synopses for a partition in bulk. Used when
   /// re-building the meta index state at startup.
   void merge(const uuid& partition, partition_synopsis&&);
-
-  /// Replaces an existing partition synopsis. Does nothing
-  /// if `partition` does not exist as a key.
-  void replace(const uuid& partition, std::unique_ptr<partition_synopsis>);
 
   /// Returns the partition synopsis for a specific partition.
   /// Note that most callers will prefer to use `lookup()` instead.
@@ -124,12 +97,5 @@ private:
   /// Settings for the synopsis factory.
   caf::settings synopsis_options_;
 };
-
-// -- flatbuffer ---------------------------------------------------------------
-
-caf::expected<flatbuffers::Offset<fbs::partition_synopsis::v0>>
-pack(flatbuffers::FlatBufferBuilder& builder, const partition_synopsis&);
-
-caf::error unpack(const fbs::partition_synopsis::v0&, partition_synopsis&);
 
 } // namespace vast
